Redesign `ContextMenuRegion` over `Menu` button to contain `MyUser` accounts Fixes #1254
