+++ lib/ui/page/home/account_switcher_menu/view.dart
@@ -1,5 +1,3 @@
-import 'dart:math' as math;
-import 'dart:ui';
 import 'dart:ui' as ui;
 
 import 'package:flutter/material.dart';
@@ -23,205 +21,6 @@ import '../tab/chats/widget/unread_counter.dart';
 import '../widget/avatar.dart';
 import '../widget/contact_tile.dart';
 import 'controller.dart';
-
-class AccountSwitcherMenuWidget extends StatefulWidget {
-  const AccountSwitcherMenuWidget({super.key, required this.child});
-
-  /// widget that viewed in layout
-  final Widget child;
-
-  @override
-  State<AccountSwitcherMenuWidget> createState() =>
-      _AccountSwitcherMenuWidgetState();
-}
-
-class _AccountSwitcherMenuWidgetState extends State<AccountSwitcherMenuWidget>
-    with SingleTickerProviderStateMixin {
-  late final OverlayPortalController _overlayController;
-
-  late final AnimationController _animationController;
-
-  @override
-  initState() {
-    _overlayController = OverlayPortalController();
-    _animationController = AnimationController(
-      duration: 150.milliseconds,
-      vsync: this,
-    );
-
-    super.initState();
-  }
-
-  @override
-  void dispose() {
-    _animationController.dispose();
-
-    super.dispose();
-  }
-
-  void _show() {
-    _overlayController.show();
-    _animationController
-      ..forward()
-      ..animateTo(1);
-  }
-
-  void _animationStatus(AnimationStatus status) {
-    if (!status.isForwardOrCompleted) return;
-    _overlayController.hide();
-
-    _animationController.removeStatusListener(_animationStatus);
-  }
-
-  void _hide() {
-    _animationController
-      ..animateTo(0)
-      ..addStatusListener(_animationStatus);
-  }
-
-  final GlobalKey _portalKey = GlobalKey();
-
-  @override
-  Widget build(BuildContext context) {
-    return OverlayPortal(
-      key: _portalKey,
-      controller: _overlayController,
-      overlayChildBuilder: _buildOverlay,
-      child: GestureDetector(
-        onLongPress: _show,
-        onSecondaryTap: _show,
-        child: widget.child,
-      ),
-    );
-  }
-
-  Widget _buildOverlay(BuildContext context) {
-    final render = _portalKey.currentContext?.findRenderObject() as RenderBox;
-
-    final portalOffset = render.localToGlobal(Offset.zero);
-
-    final portalSize = render.size;
-
-    final portalCenter = portalSize.center(portalOffset);
-
-    print([portalOffset, portalSize]);
-    print(widget.child);
-    var style = Theme.of(context).style;
-    return LayoutBuilder(
-      builder: (context, constraints) {
-        return GestureDetector(
-          behavior: HitTestBehavior.opaque,
-          onTap: _hide,
-          child: Stack(
-            children: [
-              AnimatedBuilder(
-                animation: _animationController,
-                builder: (context, child) {
-                  return FadeTransition(
-                    opacity: _animationController,
-                    child: child!,
-                  );
-                },
-                child: BackdropFilter(
-                  filter: ImageFilter.blur(sigmaX: 2, sigmaY: 2),
-                  child: Container(
-                    color: Colors.black.withValues(
-                      alpha: 0.1,
-                    ), // Semi-transparent overlay
-                  ),
-                ),
-              ),
-              AnimatedBuilder(
-                animation: _animationController,
-                builder: (context, child) {
-                  var bottom = (constraints.maxHeight - portalOffset.dy + 12);
-                  return Positioned(
-                    bottom: bottom,
-                    right:
-                        constraints.maxWidth -
-                        portalOffset.dx -
-                        portalSize.width -
-                        2,
-                    left: math.max(
-                      math.min(12, constraints.maxWidth - 500),
-                      12,
-                    ),
-                    child: FadeTransition(
-                      opacity: _animationController,
-                      child: child!,
-                    ),
-                  );
-                },
-
-                child: ConstrainedBox(
-                  constraints: BoxConstraints(
-                    minWidth: math.min(MediaQuery.sizeOf(context).width, 320),
-                  ),
-                  child: GestureDetector(
-                    behavior: HitTestBehavior.deferToChild,
-                    // must be empty for cancel propagation click event
-                    onTap: () {},
-                    child: Container(
-                      decoration: BoxDecoration(
-                        color: style.colors.background,
-                        borderRadius: BorderRadius.circular(16),
-                      ),
-                      padding: EdgeInsets.all(10),
-                      child: Column(
-                        crossAxisAlignment: CrossAxisAlignment.stretch,
-                        children: [AccountSwitcherMenuView()],
-                      ),
-                    ),
-                  ),
-                ),
-              ),
-
-              AnimatedBuilder(
-                animation: _animationController,
-                builder: (context, child) {
-                  return Positioned.fromRect(
-                    rect: Rect.fromCircle(
-                      center: portalCenter,
-                      radius: portalSize.width / 2 + 4,
-                    ),
-                    // left: portalOffset.dx - 5,
-                    // top: portalOffset.dy - 5,
-                    child: ScaleTransition(
-                      scale: Tween<double>(
-                        begin: 1,
-                        end: 1,
-                      ).animate(_animationController),
-                      alignment: Alignment.center,
-                      child: FadeTransition(
-                        opacity: Tween<double>(
-                          begin: 0,
-                          end: 1,
-                        ).animate(_animationController),
-                        child: child,
-                      ),
-                    ),
-                  );
-                },
-                child: Container(
-                  decoration: BoxDecoration(
-                    color: Colors.white,
-                    shape: BoxShape.circle,
-                  ),
-
-                  alignment: Alignment.center,
-                  child: Padding(
-                    padding: const EdgeInsets.all(1),
-                    child: widget.child,
-                  ),
-                ),
-              ),
-            ],
-          ),
-        );
-      },
-    );
-  }
-}
 
 class AccountSwitcherMenuView extends StatelessWidget {
   const AccountSwitcherMenuView({super.key});
@@ -275,9 +74,6 @@ class AccountSwitcherMenuView extends StatelessWidget {
                           await c.switchTo(myUser.id);
                         }
                       },
-
-                // TODO: Remove, when [MyUser]s will receive their
-                //       updates in real-time.
                 avatarBuilder: (_) => AvatarWidget.fromMyUser(
                   myUser,
                   radius: AvatarRadius.medium,
